feat: newsletter app
The newsletter app adds a banner form to the site, requesting an email
address to be submitted. This will then be used to send out a
subscription email in plain and html formats. Containing a confirmation
url which contains an encrypted token that is used to identify the email
address as confirmed.

This makes it possible to start building a newsletter service to send
out to the mailing list, of articles on the website.